tests: refactor converter fork tests
